+++ packages/ui/src/symbol/en.ts
@@ -213,6 +213,10 @@ export default [
         content: 'Ignore Duplicates'
     },
     {
+        name: 'upload.deduplicate',
+        content: 'Deduplicate'
+    },
+    {
         name: 'upload.dropToUpload',
         content: 'drop to upload'
     },
@@ -417,6 +421,14 @@ export default [
         content: 'addUnregistered'
     },
     {
+        name: 'entries.load.error',
+        content: 'Entries could not be loaded'
+    },
+    {
+        name: 'entry.load.error',
+        content: 'The entry could not be loaded'
+    },
+    {
         name: 'entry.select.delete.question',
         content: 'Should the entry also be deleted?'
     },
@@ -431,5 +443,17 @@ export default [
     {
         name: 'entry.select.placeholder.model',
         content: 'select model...'
+    },
+    {
+        name: 'list.empty',
+        content: 'No Item yet..'
+    },
+    {
+        name: 'list.filter.noResults',
+        content: 'No Items found'
+    },
+    {
+        name: 'list.filter.reset',
+        content: 'Clear filter'
     }
 ];